requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-registration==2.2
django-richtextfield==1.6.1
Pillow==10.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
